# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
  controllers.LoginData:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      role:
        example: admin
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      password:
        example: admin123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  controllers.LoginResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: Login successful
        type: string
    type: object
...
host: 39.108.49.167:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: intercom_http_service API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Process user login and return JWT token with different permissions
        based on user role
      parameters:
      - description: Login request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with token
          schema:
            allOf:
            - $ref: '#/definitions/controllers.LoginResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginData'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: User Login
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"